+++ openharmony/arkui_ace_engine/frameworks/core/components_ng/pattern/shape/shape_model_ng.cpp
@@ -272,7 +272,7 @@ void ShapeModelNG::SetStrokeDashArray(FrameNode* frameNode, const std::vector<Ac
             return;
         }
         std::vector<Ace::Dimension> result;
-        for (int32_t i = 0; i < segments.size(); i++) {
+        for (size_t i = 0; i < segments.size(); i++) {
             if (resObjArray[i]) {
                 Dimension dim;
                 ResourceParseUtils::ConvertFromResObjNG(resObjArray[i], dim);
